refactor(Explorer): [#471] move logic to explorer file
