+++ src/explorers/explorer.ts
@@ -6,9 +6,22 @@ import { getText } from '../domain/i18n/useCases';
 import { isTestChain, SupportedChains } from '../constants/blockchains';
 import { TransactionData } from '../models/TransactionData';
 import { ExplorerAPI } from '../certificate';
+import * as EtherscanApi from './ethereum/etherscan';
+import * as BlockExplorerApi from './bitcoin/blockexplorer';
+import * as BlockstreamApi from './bitcoin/blockstream';
+import * as BlockCypherApi from './bitcoin/blockcypher';
+import * as BitPayApi from './bitcoin/bitpay';
 
+export type TExplorerFunctionsArray = {(transactionId: string, chain: SupportedChains): Promise<TransactionData>}[];
 export type TExplorerParsingFunction = {(jsonResponse, chain?: SupportedChains): TransactionData} |
   {(jsonResponse, chain?: SupportedChains): Promise<TransactionData>};
+
+function explorerFactory (TransactionAPIArray: ExplorerAPI[]): TExplorerFunctionsArray {
+  return TransactionAPIArray
+    .map(explorerAPI =>
+      (transactionId, chain) => getTransactionFromApi(explorerAPI, transactionId, chain)
+    );
+}
 
 export async function getTransactionFromApi (
   explorerAPI: ExplorerAPI,
@@ -28,3 +41,20 @@ export async function getTransactionFromApi (
     throw new VerifierError(SUB_STEPS.fetchRemoteHash, getText('errors', 'unableToGetRemoteHash'));
   }
 }
+
+const BitcoinTransactionAPIArray = [
+  BlockCypherApi,
+  BitPayApi,
+  BlockExplorerApi,
+  BlockstreamApi
+];
+
+const EthereumTransactionAPIArray = [
+  EtherscanApi
+];
+
+export const BitcoinExplorers: TExplorerFunctionsArray = explorerFactory(BitcoinTransactionAPIArray);
+export const EthereumExplorers: TExplorerFunctionsArray = explorerFactory(EthereumTransactionAPIArray);
+
+// for legacy (pre-v2) Blockcerts
+export const BlockchainExplorersWithSpentOutputInfo: TExplorerFunctionsArray = explorerFactory([BlockCypherApi]);
